# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My docs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/arm/approveDistribution": {
            "get": {
                "tags": [
                    "ARM Distribution"
                ],
                "summary": "Получение итогового распределения",
                "description": "Контроллер получает данные о кандидатах и подходящих для них проектах на основе существующих распределений. Структура ответа включает институты, кафедры, проекты и кандидатов с массивом подходящих проектных ID.",
                "operationId": "getCandidateProject",
                "responses": {
                    "200": {
                        "description": "Успешное получение данных о кандидатах и подходящих проектах",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "projects": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "institute_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "institute_name": {
                                                        "type": "string",
                                                        "example": "Институт высоких технологий"
                                                    },
                                                    "departments": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "department_id": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "department_name": {
                                                                    "type": "string",
                                                                    "example": "Автоматизации и управления"
                                                                },
                                                                "projects": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "properties": {
                                                                            "project_id": {
                                                                                "type": "integer",
                                                                                "example": 1511
                                                                            },
                                                                            "title": {
                                                                                "type": "string",
                                                                                "example": "Разработка мобильного приложения для бронирования мест"
                                                                            },
                                                                            "places": {
                                                                                "type": "integer",
                                                                                "example": 15
                                                                            },
                                                                            "candidates_count": {
                                                                                "type": "integer",
                                                                                "example": 10
                                                                            },
                                                                            "candidates": {
                                                                                "type": "array",
                                                                                "items": {
                                                                                    "properties": {
                                                                                        "candidate_id": {
                                                                                            "type": "integer",
                                                                                            "example": 4071
                                                                                        },
                                                                                        "training_group": {
                                                                                            "type": "string",
                                                                                            "example": "АСУб-21-2"
                                                                                        },
                                                                                        "course": {
                                                                                            "type": "integer",
                                                                                            "example": 4
                                                                                        },
                                                                                        "fio": {
                                                                                            "type": "string",
                                                                                            "example": "Андреев Андрей Андреевич"
                                                                                        },
                                                                                        "created_at": {
                                                                                            "type": "string",
                                                                                            "format": "date-time",
                                                                                            "example": "2025-02-04T14:49:35.000000Z"
                                                                                        },
                                                                                        "updated_at": {
                                                                                            "type": "string",
                                                                                            "format": "date-time",
                                                                                            "example": "2025-02-04T23:37:13.000000Z"
                                                                                        },
                                                                                        "institute_id": {
                                                                                            "type": "integer",
                                                                                            "example": 4
                                                                                        },
                                                                                        "priority": {
                                                                                            "type": "integer",
                                                                                            "example": 1
                                                                                        },
                                                                                        "department_id": {
                                                                                            "type": "integer",
                                                                                            "example": 59
                                                                                        },
                                                                                        "speciality_id": {
                                                                                            "type": "integer",
                                                                                            "example": 4
                                                                                        },
                                                                                        "stranger": {
                                                                                            "type": "integer",
                                                                                            "example": 1
                                                                                        },
                                                                                        "eligible_projects_ids": {
                                                                                            "type": "array",
                                                                                            "items": {
                                                                                                "type": "integer",
                                                                                                "example": 1524
                                                                                            }
                                                                                        }
                                                                                    },
                                                                                    "type": "object"
                                                                                }
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "eligible_projects": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "project_id": {
                                                        "type": "integer",
                                                        "example": 1511
                                                    },
                                                    "project_title": {
                                                        "type": "string",
                                                        "example": "Разработка мобильного приложения"
                                                    },
                                                    "places": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "candidates_count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "specialities": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 101
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Программирование"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "institute_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "department_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при обработке запроса"
                    }
                }
            },
            "patch": {
                "tags": [
                    "ARM Distribution"
                ],
                "summary": "Обновление итогового распределения кандидатов по проектам",
                "description": "Метод используется для перемещения кандидатов между проектами. Он удаляет кандидата из одного проекта и добавляет его в другой, обновляя соответствующие данные в файле.",
                "operationId": "updateCandidateProject",
                "requestBody": {
                    "description": "Данные для обновления распределения кандидатов",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "institute_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "department_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "project_id": {
                                            "type": "integer",
                                            "example": 1001
                                        },
                                        "candidate_id": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "selected_institute_id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "selected_department_id": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "selected_project_id": {
                                            "type": "integer",
                                            "example": 2001
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление распределения кандидатов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidates transferred successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при обработке запроса"
                    },
                    "404": {
                        "description": "Файл не найден"
                    }
                }
            }
        },
...
    },
...
}